+++ src/api/routes/ProductRouter.ts
@@ -1,7 +1,7 @@
 import express from "express";
 import ProductController from "../controllers/ProductController.js";
 import authenticate from "../middlewares/authenticate.js";
-import validate from "../middlewares/validators/validate.js";
+import validate from "../middlewares/validate.js";
 import { ProductSchema } from "../middlewares/validators/validateProduct.js";
 const router = express.Router();
