+++ src/api/types/dtos/article.dto.ts
@@ -4,3 +4,5 @@ export const ArticleSchema = z.object({
   title: z.string().min(1, { message: "제목을 입력하세요" }),
   content: z.string().min(1, { message: "내용을 입력하세요." }),
 });
+
+export type ArticleDto = z.infer<typeof ArticleSchema>;
